move register commands to new file
