+++ index.js
@@ -16,17 +16,15 @@ const Path = require('path');
 const Config = require('./scripts/config');
 const Language = require('./scripts/language');
 const Discord = require('discord.js');
-const { REST } = require('@discordjs/rest');
-const { Routes } = require('discord-api-types/v9');
 
 // Local actions
 const ScriptLoader = require('./scripts/loadScripts');
+const registerInteractionCommmands = require('./scripts/registerInteractionCommands');
 const SafeMessage = require('./scripts/safeMessage');
 const CommandPermission = require('./scripts/commandPermissions');
 const MemberPermission = require('./scripts/memberPermissions');
 
 // Public vars
-const deployFile = './deploy.txt';
 const log = new Util.Logger('Bot');
 const parseConfig = new Config('./config/config.yml');
 let config = parseConfig.parse();
@@ -61,6 +59,10 @@ class UtilActions {
         commands = scriptsLoader.commands;
     }
 
+    async registerInteractionCommmands() {
+        return registerInteractionCommmands(Client, config, commands, config.guildId, false)
+    }
+
     // Commands
     async messageCommand(command, message) {
         const args = Util.getCommand(message.content.trim(), config.commandPrefix).args;
@@ -70,7 +72,7 @@ class UtilActions {
 
         // No permission
         if(!CommandPermission(command, message.member, config)) {
-            SafeMessage.reply(message, language.get(lang.noPerms));
+            SafeMessage.reply(message, Util.getRandomKey(lang.noPerms));
             return;
         }
 
@@ -79,51 +81,11 @@ class UtilActions {
         // Execute
         await scripts[command].execute(args, message, Client).catch(async err => {
             log.error(err, `${config.commandPrefix}${command}`);
-            await SafeMessage.send(message.channel, language.get(lang.error) + '\n```\n' + err.message + '\n```');
+            await SafeMessage.send(message.channel, Util.getRandomKey(lang.error) + '\n```\n' + err.message + '\n```');
         });
-    }
-    async registerInteractionCommmands(client, force = false, guild = null) {
-        // Deployment
-        if(!config.slashCommands.enabled) return;
-        if(Fs.existsSync(deployFile) && !force && !guild) {
-            const deploy = Fs.readFileSync(deployFile).toString().trim();
-
-            if(deploy == 'false') {
-                return;
-            }
-
-            Fs.writeFileSync(deployFile, 'false');
-        } else {
-            Fs.writeFileSync(deployFile, 'false'); 
-        }
-
-        // Send
-        const rest = new REST({ version: '9' }).setToken(config.token);
-        (async () => {
-            try {
-                if(!guild){
-                    await rest.put(
-                        Routes.applicationCommands(client.user.id),
-                        { body: commands },
-                    );
-                    log.warn(`${ Object.keys(commands).length } application commands were successfully registered on a global scale.`, 'Register Commands');
-                } else {
-                    await rest.put(
-                        Routes.applicationGuildCommands(client.user.id, guild),
-                        { body: commands },
-                    );
-                    log.warn(`${ Object.keys(commands).length } application commands were successfully registered on a guild.`, 'Register Commands');
-                }
-            } catch (err) {
-                log.error(err, 'Register Commands');
-            }
-        })();
     }
 
     // Other utility functions
-    get(object) {
-        return language.get(object);
-    }
     createInvite(bot) {
         return Util.replaceAll(config.inviteFormat, '%id%', bot.user.id);
     }
@@ -139,7 +101,7 @@ Client.once('ready', async () => {
     
     // Register commands
     await Actions.loadScripts();
-    await Actions.registerInteractionCommmands(Client, false, config.guildId);
+    await Actions.registerInteractionCommmands();
 });
 
 Client.on('ready', function() {
@@ -154,14 +116,14 @@ Client.on('ready', function() {
         // Check configurations
         if(!config.slashCommands.enabled || MemberPermission.isIgnoredChannel(interaction.channelId, config)) { 
             await interaction.reply({ 
-                content: language.get(lang.notAvailable),
+                content: Util.getRandomKey(lang.notAvailable),
                 ephemeral: true
             }).catch(err => log.error(err));
             return; 
         }
         if(!CommandPermission(command['command']['name'], interaction.member, config)) { 
             interaction.reply({ 
-                content: language.get(lang.noPerms),
+                content: Util.getRandomKey(lang.noPerms),
                 ephemeral: true
             }).catch(err => log.error(err, 'Slash command'));
             return;
